+++ components/bridge/synchronization/impl/src/main/java/com/flipperdevices/bridge/synchronization/impl/repository/android/AndroidHashRepository.kt
@@ -1,12 +1,20 @@
 package com.flipperdevices.bridge.synchronization.impl.repository.android
 
 import com.flipperdevices.bridge.dao.api.model.FlipperKey
+import com.flipperdevices.bridge.synchronization.impl.di.TaskGraph
 import com.flipperdevices.bridge.synchronization.impl.model.KeyWithHash
 import com.flipperdevices.core.ktx.jre.md5
 import com.flipperdevices.core.ktx.jre.pmap
+import com.squareup.anvil.annotations.ContributesBinding
+import javax.inject.Inject
 
-class AndroidHashRepository {
-    suspend fun calculateHash(keys: List<FlipperKey>): List<KeyWithHash> {
+interface AndroidHashRepository {
+    suspend fun calculateHash(keys: List<FlipperKey>): List<KeyWithHash>
+}
+
+@ContributesBinding(TaskGraph::class, AndroidHashRepository::class)
+class AndroidHashRepositoryImpl @Inject constructor() : AndroidHashRepository {
+    override suspend fun calculateHash(keys: List<FlipperKey>): List<KeyWithHash> {
         return keys.pmap { flipperKey ->
             flipperKey.additionalFiles.plus(flipperKey.mainFile).map {
                 KeyWithHash(