+++ components/bridge/synchronization/impl/src/main/java/com/flipperdevices/bridge/synchronization/impl/repository/flipper/FlipperHashRepository.kt
@@ -4,6 +4,7 @@ import com.flipperdevices.bridge.api.manager.FlipperRequestApi
 import com.flipperdevices.bridge.api.model.FlipperRequestPriority
 import com.flipperdevices.bridge.api.model.wrapToRequest
 import com.flipperdevices.bridge.dao.api.model.FlipperFilePath
+import com.flipperdevices.bridge.synchronization.impl.di.TaskGraph
 import com.flipperdevices.bridge.synchronization.impl.model.KeyWithHash
 import com.flipperdevices.bridge.synchronization.impl.model.ResultWithProgress
 import com.flipperdevices.core.ktx.jre.pmap
@@ -11,22 +12,33 @@ import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
 import com.flipperdevices.protobuf.main
 import com.flipperdevices.protobuf.storage.md5sumRequest
+import com.squareup.anvil.annotations.ContributesBinding
 import java.util.concurrent.atomic.AtomicInteger
+import javax.inject.Inject
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.single
 
-class FlipperHashRepository : LogTagProvider {
+interface FlipperHashRepository {
+    fun calculateHash(
+        keys: List<FlipperFilePath>
+    ): Flow<ResultWithProgress<List<KeyWithHash>>>
+}
+
+@ContributesBinding(TaskGraph::class, FlipperHashRepository::class)
+class FlipperHashRepositoryImpl @Inject constructor(
+    private val requestApi: FlipperRequestApi
+) : FlipperHashRepository, LogTagProvider {
     override val TAG = "HashRepository"
 
-    fun calculateHash(
-        requestApi: FlipperRequestApi,
+    override fun calculateHash(
         keys: List<FlipperFilePath>
     ) = callbackFlow {
         val alreadyHashReceiveCounter = AtomicInteger(0)
         info { "Start request hashes for ${keys.size} keys" }
 
         val hashList = keys.pmap { keyPath ->
-            val hash = receiveHash(requestApi, keyPath)
+            val hash = receiveHash(keyPath)
             send(
                 ResultWithProgress.InProgress(
                     currentPosition = alreadyHashReceiveCounter.incrementAndGet(),
@@ -41,7 +53,6 @@ class FlipperHashRepository : LogTagProvider {
     }
 
     private suspend fun receiveHash(
-        requestApi: FlipperRequestApi,
         keyPath: FlipperFilePath
     ): String {
         return requestApi.request(