+++ components/bridge/synchronization/impl/src/main/java/com/flipperdevices/bridge/synchronization/impl/repository/flipper/KeysListingRepository.kt
@@ -7,6 +7,7 @@ import com.flipperdevices.bridge.api.utils.Constants
 import com.flipperdevices.bridge.dao.api.model.FlipperFilePath
 import com.flipperdevices.bridge.dao.api.model.FlipperFileType
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyType
+import com.flipperdevices.bridge.synchronization.impl.di.TaskGraph
 import com.flipperdevices.bridge.synchronization.impl.model.ResultWithProgress
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.debug
@@ -14,30 +15,37 @@ import com.flipperdevices.core.log.info
 import com.flipperdevices.protobuf.main
 import com.flipperdevices.protobuf.storage.Storage
 import com.flipperdevices.protobuf.storage.listRequest
+import com.squareup.anvil.annotations.ContributesBinding
 import java.io.File
+import javax.inject.Inject
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.toList
 
 const val SIZE_BYTES_LIMIT = 10 * 1024 * 1024 // 10MiB
 
-class KeysListingRepository : LogTagProvider {
+interface KeysListingRepository {
+    fun getAllKeys(): Flow<ResultWithProgress<List<FlipperFilePath>>>
+}
+
+@ContributesBinding(TaskGraph::class, KeysListingRepository::class)
+class KeysListingRepositoryImpl @Inject constructor(
+    private val requestApi: FlipperRequestApi
+) : KeysListingRepository, LogTagProvider {
     override val TAG = "KeysListingRepository"
 
-    fun getAllKeys(
-        requestApi: FlipperRequestApi
-    ) = callbackFlow<ResultWithProgress<List<FlipperFilePath>>> {
+    override fun getAllKeys() = callbackFlow<ResultWithProgress<List<FlipperFilePath>>> {
         info { "Start request keys listing" }
         val allKeys = mutableListOf<FlipperFilePath>()
         FlipperKeyType.values().forEachIndexed { index, fileType ->
             send(ResultWithProgress.InProgress(index, FlipperKeyType.values().size))
-            allKeys.addAll(getKeysForFileType(requestApi, fileType))
+            allKeys.addAll(getKeysForFileType(fileType))
         }
         send(ResultWithProgress.Completed(allKeys))
         close()
     }
 
     private suspend fun getKeysForFileType(
-        requestApi: FlipperRequestApi,
         fileType: FlipperKeyType
     ): List<FlipperFilePath> {
         val fileTypePath = File(Constants.KEYS_DEFAULT_STORAGE, fileType.flipperDir).path